HIstorial de Acceso de usuarios

+++ resources/views/usuarios/partials/table.blade.php
@@ -22,10 +22,12 @@
 					<td>{{ ($user->status == 1)?'Confirmado':'No Confirmado' }}</td>
 					<td>						
                         <div class='btn-group  btn-group-xs' role='group' aria-label='...'>
-		                    {!! Form::model($user, ["url" => action('UserController@destroy', $user->id), "method" => "DELETE"]) !!}
-							<button type="button" class="btn btn-info btn-xs"><i class="glyphicon glyphicon-link" title="Reenviar correo de confirmación"></i></button>
+		                    {!! Form::model($user, ["url" => action('UserController@getConfirmation', $user->id), "method" => "GET"]) !!}
+								<button class='btn btn-info btn-xs' type="submit"><i class="glyphicon glyphicon-link" title="Reenviar correo de confirmación"></i></button>
+							{!! Form::close() !!}
 							<a href='{{action('UserController@edit', ["id"=>$user->id])}}'  type="button" class="btn btn-warning btn-xs"><i class="glyphicon glyphicon-edit" title="Editar"></i></a>
-								<button class='btn btn-danger btn-xs' type="submit"><i class="glyphicon glyphicon-remove" title="Eliminar"></i></button>
+		                    {!! Form::model($user, ["url" => action('UserController@destroy', $user->id), "method" => "DELETE",  'onsubmit' => 'return ConfirmDelete()']) !!}
+								<button class='btn btn-danger btn-xs delete-btn' type="submit"><i class="glyphicon glyphicon-remove" title="Eliminar"></i></button>
 							{!! Form::close() !!}
 						</div>
 					</td>
